pyproject.toml
[tool.poetry]
name = 'good-first-issue'
version = '0.1.0'
description = 'https://goodfirstissue.dev'
license = 'MIT'
authors = [
    'Sanket Saurav <sanket@deepsource.io>'
]

[tool.poetry.dependencies]
python = "^3.6"
"github3.py" = "^1.3"
toml = "^0.10.0"
algoliasearch = "^2.1.0"
numerize = "^0.12"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
